refactor: Add subdirectory on pantry dto (PLTR-58)
