+++ server/plantry/src/main/java/com/gdscplantry/plantry/domain/Pantry/domain/UserPantryRepository.java
@@ -1,6 +1,6 @@
 package com.gdscplantry.plantry.domain.Pantry.domain;
 
-import com.gdscplantry.plantry.domain.Pantry.dto.PantryListItemDto;
+import com.gdscplantry.plantry.domain.Pantry.dto.pantry.PantryListItemDto;
 import com.gdscplantry.plantry.domain.User.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -11,7 +11,7 @@ import java.util.Optional;
 
 @Repository
 public interface UserPantryRepository extends JpaRepository<UserPantry, Long> {
-    @Query(value = "select new com.gdscplantry.plantry.domain.Pantry.dto.PantryListItemDto(up.pantryId, up.title, up.color, up.isMarked) " +
+    @Query(value = "select new com.gdscplantry.plantry.domain.Pantry.dto.pantry.PantryListItemDto(up.pantryId, up.title, up.color, up.isMarked) " +
             "from UserPantry up " +
             "where up.user = :user " +
             "order by up.isMarked desc, up.title asc")
