+++ server/plantry/src/main/java/com/gdscplantry/plantry/domain/Pantry/dto/pantry/NewPantryReqDto.java
@@ -1,4 +1,4 @@
-package com.gdscplantry.plantry.domain.Pantry.dto;
+package com.gdscplantry.plantry.domain.Pantry.dto.pantry;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
@@ -14,7 +14,7 @@ public class NewPantryReqDto {
     @Size(max = 12, message = "Invalid title value. (Max length is 12)")
     @NotBlank(message = "Title required.")
     private String title;
-    
+
     @Pattern(regexp = "^([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$", message = "Invalid color value.")
     private String color;
 }
