+++ server/plantry/src/main/java/com/gdscplantry/plantry/domain/Pantry/dto/pantry/PantryListResDto.java
@@ -1,4 +1,4 @@
-package com.gdscplantry.plantry.domain.Pantry.dto;
+package com.gdscplantry.plantry.domain.Pantry.dto.pantry;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
