+++ server/plantry/src/test/java/com/gdscplantry/plantry/domain/Pantry/controller/PantryControllerTest.java
@@ -5,8 +5,8 @@ import com.gdscplantry.plantry.domain.Pantry.domain.Pantry;
 import com.gdscplantry.plantry.domain.Pantry.domain.PantryRepository;
 import com.gdscplantry.plantry.domain.Pantry.domain.UserPantry;
 import com.gdscplantry.plantry.domain.Pantry.domain.UserPantryRepository;
-import com.gdscplantry.plantry.domain.Pantry.dto.NewPantryReqDto;
-import com.gdscplantry.plantry.domain.Pantry.dto.UpdatePantryReqDto;
+import com.gdscplantry.plantry.domain.Pantry.dto.pantry.NewPantryReqDto;
+import com.gdscplantry.plantry.domain.Pantry.dto.pantry.UpdatePantryReqDto;
 import com.gdscplantry.plantry.domain.User.domain.User;
 import com.gdscplantry.plantry.domain.User.domain.UserRepository;
 import com.gdscplantry.plantry.domain.model.JwtVo;
